requirements.txt
absl-py==2.0.0
ale-py==0.8.1
anyio==4.4.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
beautifulsoup4==4.12.3
bleach==6.1.0
bson==0.5.8
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.1.0
crypto==1.4.1
cycler==0.11.0
Cython==3.0.10
dacite==1.8.1
dataclasses==0.6
defusedxml==0.7.1
distlib==0.3.7
distro==1.9.0
Django==4.2.14
django-betterforms==1.2
django-bootstrap4==24.3
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-jazzmin==2.6.1
django-js-asset==1.2.2
django-khalti==1.0
django-mapbox-location-field==1.5.0
django-mathfilters==1.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==1.16.0
email_validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.12.3
findspark==2.0.1
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.59.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.10.0
htmlmin==0.1.12
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
ImageHash==4.3.1
imageio==2.34.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==2.15.0
kiwisolver==1.4.4
lazy_loader==0.4
libclang==16.0.6
lxml==5.1.1
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
multimethod==1.9.1
mysqlclient==2.2.4
Naked==0.1.32
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.30.3
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.3
pandas-profiling==3.6.6
pandocfilters==1.5.1
patsy==0.5.3
phik==0.12.3
pillow==10.2.0
pipenv==2023.9.1
platformdirs==3.10.0
pmdarima==2.0.4
pretty-confusion-matrix==0.2.0
protobuf==4.24.4
psycopg2==2.9.9
py4j==0.10.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==1.10.12
pygame==2.5.2
Pygments==2.17.2
PyJWT==2.9.0
pymunk==6.6.0
pynput==1.7.7
pyparsing==3.0.9
pyperclip==1.9.0
PySocks==1.7.1
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyTMX==3.32
pytz==2020.1
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.11.2
seaborn==0.13.2
selenium==4.21.0
shellescape==3.8.1
shellingham==1.5.4
six==1.15.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
starlette==0.37.2
statsmodels==0.14.0
tangled-up-in-unicode==0.2.0
tensorboard==2.15.2
tensorboard-data-server==0.7.1
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2024.5.10
tinycss2==1.3.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typeguard==2.13.3
typer==0.12.3
typing_extensions==4.12.0
tzdata==2023.3
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
virtualenv==20.24.4
visions==0.7.5
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.0
wordcloud==1.9.2
wrapt==1.14.1
wsproto==1.2.0
ydata-profiling==4.5.1
zipp==3.16.2
